refactor(cms-plugin): split translation styles
