+++ libs/cms-plugin/src/components/client/translations-select-locales-drawer.module.css
@@ -1,0 +1,19 @@
+.selectLocalesText {
+  margin-top: 2rem;
+}
+
+.selectLocalesList {
+  margin-top: 2rem;
+
+  > div {
+    margin-top: 1rem;
+  }
+}
+
+.selectLocalesNote {
+  margin-top: 2rem;
+}
+
+.selectLocalesFooter {
+  margin-top: 0.25rem;
+}
